+++ homeclouds/nb-configuration.xml
