+++ homeclouds/src/main/java/support/AbsPath.java
@@ -1,0 +1,5 @@
+package support;
+
+public class AbsPath {
+    static String path = "D:\\University\\Cloud\\";
+}
