+++ homeclouds/src/main/java/trying_db2/DatabaseFunction.java
