--- mavenproject1/src/main/java/support/AbsPath.java
@@ -1,3 +1,0 @@
-public class AbsPath {
-    static String path = "D:\\University\\Cloud\\";
-}
