--- mavenproject1/src/main/webapp/WEB-INF/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <title>HomeClouds</title>
-    </head>
-    <body>
-        <h1>Home page</h1>
-        <a href="user" > User page </a><br>
-        <a href="auth">Sign in</a>
-        <p>New user? <a href="register">Sign up</a></p>
-    </body>
-</html>
-
