--- mavenproject1/src/main/webapp/WEB-INF/web.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-     xmlns="http://java.sun.com/xml/ns/javaee"
-     xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-                       http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-     id="WebApp_ID" version="3.0">
- <display-name>SimpleWebApp</display-name>
- 
- <welcome-file-list>
- 
-   <welcome-file>/</welcome-file>
-    
-   <welcome-file>index.html</welcome-file>
-    
- </welcome-file-list>
- 
- 
-</web-app>
